OCM-3677 | fix: cluster-autoscaler enabled for HCP
When creating a HCP cluster, or when handling it in day2 operations,
`rosa`` CLI allows handling of cluster-autoscaler resource even though
it's not yet supported.

This displays an error and exits the program for those cases.

+++ cmd/dlt/autoscaler/cmd.go
@@ -46,6 +46,11 @@ func run(cmd *cobra.Command, _ []string) {
 	clusterKey := r.GetClusterKey()
 	cluster := r.FetchCluster()
 
+	if cluster.Hypershift().Enabled() {
+		r.Reporter.Errorf("Hosted Control Plane clusters do not support cluster-autoscaler configuration")
+		os.Exit(1)
+	}
+
 	r.Reporter.Debugf("Deleting autoscaler for cluster '%s''", clusterKey)
 
 	err := r.OCMClient.DeleteClusterAutoscaler(cluster.ID())
